Refactorization repository Specialty

+++ src/main/java/com/hegaro/medicalapp/repository/SpecialtyRepository.java
@@ -3,5 +3,5 @@ package com.hegaro.medicalapp.repository;
 import com.hegaro.medicalapp.model.Specialty;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpecialtyRepository extends JpaRepository<Specialty, Integer> {
+public interface SpecialtyRepository extends JpaRepository<Specialty, Long> {
 }
